requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.46
databases==0.6.1
jinja2==3.1.2
python-multipart==0.0.6
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
email-validator==1.3.1
python-jose==3.3.0
python-dotenv==1.0.0
aiofiles==23.1.0
csvkit==1.0.7
argon2-cffi==21.3.0